+++ src/components/StatusIcon/StatusIcon.tsx
@@ -169,8 +169,6 @@ export function StatusIcon({
         width={size}
         xmlns="http://www.w3.org/2000/svg"
       >
-        <title id="svg-title">{setAriaLabel}</title>
-        <desc id="svg-description">{setAriaDesc}</desc>
         {(!level || level < 0 || level > 5) && (
           <g>
             <path d="" transform="translate(-45 -15)" role="presentation" />